+++ src/ui/mod.rs
@@ -6,6 +6,7 @@ pub mod context_menu;
 pub mod discovery_handler;
 pub mod dummy_data;
 pub mod header_bar;
+mod library_commands;
 pub mod objects;
 pub mod open_files;
 pub mod output_dialogs;